Discretizer rewrite

+++ Discretizer/SRAD_Discretizer.py
@@ -1,26 +1,32 @@
 import csv
 import os
+import subprocess
 from itertools import product, chain
+from shutil import which
+
 from pandas import DataFrame
 
 import pandas as pd
 from pgmpy.models import BayesianModel
 
-from Discretizer import DiscretizerFactory
+from Discretizer import DiscretizerFactory, Discretizer
 from Discretizer.AbstractSupervisedDiscretizer import AbstractSupervisedDiscretizer
 from typing import NewType, Dict, List, Tuple
+from gobnilp import parse_gobnilp_structure
 
 HyperParameter = NewType("HyperParameter", int)
 Grid = NewType("Grid", Dict[str, List[HyperParameter]])
 
 
 class SRAD_Discretizer(AbstractSupervisedDiscretizer):
+    grid = {'EWD': [2, 3, 10], 'EFD': [2, 3, 10], 'IQR': [0], 'Median': [0]}
 
     def __init__(self, parent_limit=6, alpha=10,
-                 grid={'EWD': [2, 3, 10], 'EFD': [2, 3, 10], 'IQR': [0], 'Median': [0]}):
+                 grid=None):
         self.parent_limit = parent_limit
         self.alpha = alpha
-        self.grid = grid
+        if grid:
+            self.grid = grid
 
     @staticmethod
     def unfold_grid(grid: object) -> List[Tuple[str, int]]:
@@ -47,20 +53,20 @@ class SRAD_Discretizer(AbstractSupervisedDiscretizer):
         dff.loc[-1] = [len(pd.Series(dff[x].unique())) for x in dff.columns]
         dff.index = dff.index + 1
         dff = dff.sort_index()
-
+        print(dff)
         return dff.to_csv(sep=' ', index=False, quoting=csv.QUOTE_NONE)
 
-    def get_discretization_graph(self, df: DataFrame, discretization_var: str, grid: Grid) -> BayesianModel:
+    def get_discretization_graph(self, df: DataFrame, discretization_var: str, grid: Grid,
+                                 objective: str) -> BayesianModel:
         ddf = self.create_discretization_dataframe(df, discretization_var, grid)
-        constraints = self.create_constraints([discretization_var], grid)
+        constraints = self.create_constraints([discretization_var], grid, objective)
         G = self.structure_learn(self.discr_df_to_dat(ddf), self.get_discretization_settings(), constraints)
         return G
 
-    @classmethod
-    def get_discretization_strategy(cls, df: DataFrame, discretization_var: str, objective: str, grid: Grid) -> List[
+    def get_discretization_strategy(self, df: DataFrame, discretization_var: str, objective: str, grid: Grid) -> List[
         str]:
-        G = cls.get_discretization_graph(df, discretization_var, grid)
-        selected_nodes = [cls.from_strategy_name(x) for (x, y) in G.edges if
+        G = self.get_discretization_graph(df, discretization_var, grid, objective)
+        selected_nodes = [self.from_strategy_name(x) for (x, y) in G.edges() if
                           y == objective and discretization_var in x]
         return selected_nodes
 
@@ -68,18 +74,19 @@ class SRAD_Discretizer(AbstractSupervisedDiscretizer):
     @classmethod
     def create_discretization_dataframe(cls, df: DataFrame, discretization_var: str, grid: Grid) -> DataFrame:
         ddf = df.copy()
-        ddf = ddf.drop(discretization_var, axis=1)
+        # ddf = ddf.drop(discretization_var, axis=1)
 
         strategies = []
         for method in grid:
             for hp in grid[method]:
                 strategies.append((discretization_var, method, hp))
 
-        return cls.discretize_from_strategies(ddf, strategies)
+        print(strategies)
+        return Discretizer.discretize_from_strategies(ddf, strategies)
 
-    @classmethod
-    def get_raw_bins(cls, column: pd.Series, target: str, df: DataFrame) -> List[int]:
-        selected_nodes = cls.get_discretization_strategy(df, column.name, target, cls.grid)
+    def get_raw_bins(self, column: pd.Series, target: str, df: DataFrame) -> List[int]:
+        print(df)
+        selected_nodes = self.get_discretization_strategy(df, column.iloc[:, 0].name, target, self.grid)
 
         # In this version we're going to always select the first strategy, and we assume only one exists..
         discretization_var, method, hp = selected_nodes[0]
@@ -88,14 +95,32 @@ class SRAD_Discretizer(AbstractSupervisedDiscretizer):
         if isinstance(discretizer, AbstractSupervisedDiscretizer):
             raise ValueError("Using supervised discretizers has not yet been implemented for SRAD")
 
-        return discretizer.get_bins(hp)
-        
+        return discretizer.get_bins(df[discretization_var],hp)
+
     @classmethod
     def get_name(cls):
         return "SRAD"
 
     @staticmethod
-    def structure_learn(dat, settings, constraints):
+    def structure_learn(dat: str, settings: str, constraints: List[str]) -> object:
+
+        if not which('gobnilp'):
+            raise EnvironmentError("Could not find gobnilp on path; make sure it is installed.")
 
         if os.path.exists('sol.dot'):
             os.remove('sol.dot')
+
+        with open('datafile.dat', 'w') as data_file, open('gobnilp.set', 'w') as settings_file, open('constraints.con',
+                                                                                                     'w') as constraints_file:
+            data_file.write(dat)
+            settings_file.write(settings)
+            constraints_file.write('\n'.join(constraints))
+
+        proc = subprocess.Popen(['gobnilp', 'datafile.dat'], stdout=subprocess.PIPE, stderr=subprocess.PIPE,
+                                   cwd='.')
+        proc.wait()
+        out, err = proc.communicate()
+        print(err)
+
+        solution = open('sol.dot', 'r')
+        return parse_gobnilp_structure(solution)
